test(server): split core setup tests into focused suites
